refactor(FacebookAuth): move production code to their own files
