+++ src/infra/http/client.ts
@@ -1,0 +1,11 @@
+
+export interface HttpGetClient {
+  get: (params: HttpGetClient.Params) => Promise<void>
+}
+
+export namespace HttpGetClient {
+  export type Params = {
+    url: string
+    params: object
+  }
+}
